fix: Conversion from long double to unsigned long long
The problem showed up in the A_matrix::num_p_stars(...) method for counting 3-stars in a network with 5 nodes in which case the maximum number of 3-stars per node is 4, but the conversion of 4.000000 to unsigned long long gives 3 on both clang++ and g++. Now the functions just return long double as aux_math::binom(...) does to avoid the conversion.

+++ src/matrices/A_matrix.cpp
@@ -67,19 +67,19 @@ col_vector<unsigned int> A_matrix::degree_sequence_col_vec() const {
   return deg_seq;
 }
 
-unsigned long long A_matrix::num_p_stars(const unsigned int& i, const unsigned int& p) const {
+long double A_matrix::num_p_stars(const unsigned int& i, const unsigned int& p) const {
   return aux_math::binom(degree(i), p);
 }
 
-col_vector<unsigned long long> A_matrix::num_p_stars_sequence_col_vec(const unsigned int& p) const {
-  col_vector<unsigned long long> p_stars_sequence(dim_x);
+col_vector<long double> A_matrix::num_p_stars_sequence_col_vec(const unsigned int& p) const {
+  col_vector<long double> p_stars_sequence(dim_x);
   for(unsigned int i=0; i<dim_x; ++i)
     p_stars_sequence[i] = num_p_stars(i, p);
   return p_stars_sequence;
 }
 
-double A_matrix::average_p_stars(const unsigned int& p) const {
-  double s = 0;
+long double A_matrix::average_p_stars(const unsigned int& p) const {
+  long double s = 0;
   for(unsigned int i=0; i<dim_x; ++i)
     s+=num_p_stars(i, p);
   return s/dim_x;
