+++ src/matrices/A_matrix.hpp
@@ -40,9 +40,9 @@ class A_matrix : public symm_matrix<bool> {//Adjacency matrix
   unsigned int num_links() const;
   unsigned int degree(const unsigned int& i) const;
   col_vector<unsigned int> degree_sequence_col_vec() const;
-  col_vector<unsigned long long> num_p_stars_sequence_col_vec(const unsigned int& p) const;
-  double average_p_stars(const unsigned int& p) const;
-  unsigned long long num_p_stars(const unsigned int& node, const unsigned int& p) const;
+  col_vector<long double> num_p_stars_sequence_col_vec(const unsigned int& p) const;
+  long double average_p_stars(const unsigned int& p) const;
+  long double num_p_stars(const unsigned int& node, const unsigned int& p) const;
   unsigned long long num_triangles(const unsigned int& node) const; // Number of triangles with a given node
   unsigned long long num_triangles() const; // Full number of triangles in the network (INEFFICIENT!!!)
   double loc_clust_coeff(const unsigned int& node) const;
